Issue #1792 Moved recipients logic to ComposerContacts.ts file
